added fiware-monitoring

+++ deployment/puppet/context-broker/files/install_libmicrohttpd.sh
@@ -3,4 +3,6 @@ cd /home/context-broker/libmicrohttpd-0.9.16/
 ./configure
 make
 make install
-ln -s /usr/local/lib/libmicrohttpd.so.10 /usr/lib/libmicrohttpd.so.10
+if [ ! -L 'usr/lib/libmicrohttpd.so.10' ]; then
+	ln -s /usr/local/lib/libmicrohttpd.so.10 /usr/lib/libmicrohttpd.so.10
+fi
